+++ API/test/test_views.py
@@ -1,4 +1,3 @@
-import os
 import pytest
 from django.test import override_settings
 from django.core.files.uploadedfile import SimpleUploadedFile
@@ -6,18 +5,12 @@ from django.urls import reverse
 from rest_framework.authtoken.models import Token
 from rest_framework.test import APIClient
 from API.models import CustomThumbnailSize, AccountTypePermissions, APIUserProfile, StoredImage, GeneratedImage
-from ImageUploadAPI.settings import TEST_API_DIR
 from API.test.constants_tests import TEST_USER_PASS, TEST_USER_LOGIN, TEST_MEDIA_ROOT, TEST_MEDIA_URL, \
     TEST_PROFILE_TYPE_NAME
 from API.test.utils import db_data_preparation
 
+
 pytestmark = pytest.mark.django_db  # all test functions can access db
-
-
-@pytest.fixture(scope="session", autouse=True)
-def directory_setup():
-    """Move to /API/test directory  before tests start, to ensure access to test image file"""
-    os.chdir(TEST_API_DIR)
 
 
 @override_settings(MEDIA_URL=TEST_MEDIA_URL, MEDIA_ROOT=TEST_MEDIA_ROOT)
